change objective to scenario tab and move global objective to model tab
